requirements.txt
# --- Framework API ---
fastapi[all]
uvicorn

# --- Traitement de données ---
pandas
numpy

# --- Machine Learning & NLP ---
scikit-learn
sentence-transformers
torch
joblib

# --- AWS & Infrastructure ---
boto3

# --- Application UI ---
streamlit
requests

# --- Gestion des variables d'environnement ---
python-dotenv